==> notebooks/01_download_datasets.ipynb <==
import os

path = "../data/raw/"
if not os.path.exists(path):
    os.makedirs(path)

%cd ../data/raw/

# Initial human abstracts
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-init.jsonl
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-init.xlsx

# Abstracts fully generated by ChatGPT
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-chatgpt-generation.jsonl
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-chatgpt-generation.xlsx

# Abstracts polished by ChatGPT
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-chatgpt-polish.jsonl
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-chatgpt-polish.xlsx

# Mixed abstracts (Fusion)
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-chatgpt-fusion.jsonl
!wget https://raw.githubusercontent.com/botianzhe/CHEAT/main/data/ieee-chatgpt-fusion.xlsx
