==> notebooks/02_train_test_splitting.ipynb <==
# !pip install pandas
# !pip install scikit-learn

import pandas as pd
from sklearn.model_selection import train_test_split

# Label dictionary → path to the JSONL
raw_data_path = '../data/raw/'
files = {
    'human': f'{raw_data_path}ieee-init.jsonl',
    'generation': f'{raw_data_path}ieee-chatgpt-generation.jsonl',
    'polish': f'{raw_data_path}ieee-chatgpt-polish.jsonl',
    'mix': f'{raw_data_path}ieee-chatgpt-fusion.jsonl'
}

df_human = pd.read_json(files['human'], lines=True)
df_human.head(5)

# Split parameters
split_params = {
    'test_size': 0.2,
    'random_state': 42,
    'shuffle': True
}

# Split the data
dfs_train, dfs_test = [], []
for label, file in files.items():
    # Read the JSONL file
    df = pd.read_json(file, lines=True)[['abstract']]
    df['label'] = label # Add the label column
    # Split the data into train and test sets
    df_train, df_test = train_test_split(df, **split_params)
    # Save the train and test sets to CSV files
    dfs_train.append(df_train)
    dfs_test.append(df_test)

df_train = pd.concat(dfs_train, ignore_index=True)
df_test = pd.concat(dfs_test, ignore_index=True)

# Save the train and test sets to CSV files
save_path = '../data/train_test/'
df_train.to_csv(f'{save_path}train.csv', index=False)
df_test.to_csv(f'{save_path}test.csv', index=False)
